Fix: use float for credits to support 3.5 credit courses

+++ app.py
@@ -31,7 +31,6 @@ GRADE_CONVERSION = {
 
 def load():
     # Read all courses from the CSV file and return as a list of dictionaries
-    # Each row becomes a dict like: {"id": "1", "name": "COMP 232", ...}
     # If the file doesn't exist yet, return an empty list
     if not os.path.exists(FILE):
         return []
@@ -41,40 +40,33 @@ def load():
 
 def save(courses):
     # Write the full list of courses back to the CSV file
-    # This completely overwrites the file each time — simple but effective for small data
     with open(FILE, "w", newline="") as f:
         writer = csv.DictWriter(f, fieldnames=HEADERS)
-        writer.writeheader()   # Write the column names first
-        writer.writerows(courses)  # Then write all course rows
+        writer.writeheader()
+        writer.writerows(courses)
 
 
 def next_id(courses):
-    # Generate the next unique ID by finding the current highest ID and adding 1
-    # If no courses exist yet, start at ID 1
+    # Generate the next unique ID
     return str(max(int(c["id"]) for c in courses) + 1) if courses else "1"
 
 
 def calculate_cgpa(courses):
     # Calculate CGPA using Concordia's weighted average formula:
-    # CGPA = sum(credit_weight * gpa_points) / sum(credit_weights)
-    # Only courses with a grade are included — in-progress courses are excluded
-    graded = [(int(c["credits"]), float(c["grade"])) for c in courses if c["grade"]]
+    # CGPA = sum(credits * gpa_points) / sum(credits)
+    # Use float for credits to support 3.5 credit courses
+    graded = [(float(c["credits"]), float(c["grade"])) for c in courses if c["grade"]]
     if not graded:
-        return None  # No grades yet, nothing to calculate
-    total_points = sum(c * g for c, g in graded)   # Sum of (credits * gpa) for each course
-    total_credits = sum(c for c, g in graded)       # Sum of all graded credits
+        return None
+    total_points = sum(c * g for c, g in graded)
+    total_credits = sum(c for c, g in graded)
     return round(total_points / total_credits, 2)
 
 
 def target_needed(courses, target, upcoming):
-    # Calculate what average GPA you need in your upcoming credits
-    # to reach your target CGPA overall
-    #
-    # Formula derived from the CGPA equation:
-    # target = (current_points + upcoming_credits * x) / (current_credits + upcoming_credits)
-    # Solving for x:
-    # x = (target * total_future_credits - current_points) / upcoming_credits
-    graded = [(int(c["credits"]), float(c["grade"])) for c in courses if c["grade"]]
+    # Calculate what average GPA you need in upcoming credits to reach target CGPA
+    # Use float for credits to support 3.5 credit courses
+    graded = [(float(c["credits"]), float(c["grade"])) for c in courses if c["grade"]]
     total_points = sum(c * g for c, g in graded)
     total_credits = sum(c for c, g in graded)
     return round((target * (total_credits + upcoming) - total_points) / upcoming, 2)
@@ -85,38 +77,31 @@ def target_needed(courses, target, upcoming):
 def index():
     courses = load()
     cgpa = calculate_cgpa(courses)
-    # Build the list of grade options for the dropdown (exclude the empty placeholder)
     grade_options = [k for k in GRADE_CONVERSION.keys() if k != ""]
     return render_template("index.html", courses=courses, cgpa=cgpa, grade_options=grade_options)
 
 
 # Handle form submission to add a new course
-# POST only — this route should never be accessed directly via URL
 @app.route("/add", methods=["POST"])
 def add():
     courses = load()
-    # Get the selected letter grade from the form and convert it to a GPA value
-    # e.g. "B+" becomes 3.3
+    # Convert letter grade to GPA value before saving
     letter = request.form["grade"]
     gpa_value = GRADE_CONVERSION.get(letter, None)
-    # Append the new course to the list
     courses.append({
         "id": next_id(courses),
         "name": request.form["name"],
         "credits": request.form["credits"],
-        "grade": gpa_value if gpa_value is not None else "",  # Store empty string if no grade
+        "grade": gpa_value if gpa_value is not None else "",
         "semester": request.form["semester"]
     })
     save(courses)
-    # Redirect back to home page so the user sees the updated list
     return redirect("/")
 
 
 # Handle deleting a course by its ID
-# The ID is passed directly in the URL e.g. /delete/3
 @app.route("/delete/<id>")
 def delete(id):
-    # Rebuild the list excluding the course with the matching ID
     courses = [c for c in load() if c["id"] != id]
     save(courses)
     return redirect("/")
@@ -126,17 +111,15 @@ def delete(id):
 @app.route("/target", methods=["POST"])
 def target():
     courses = load()
-    target = float(request.form["target"])       # Desired CGPA
-    upcoming = int(request.form["upcoming"])     # Credits still to be completed
+    target = float(request.form["target"])
+    upcoming = int(request.form["upcoming"])
     cgpa = calculate_cgpa(courses)
     needed = target_needed(courses, target, upcoming)
     grade_options = [k for k in GRADE_CONVERSION.keys() if k != ""]
-    # Pass the result back to the template to display the message
     return render_template("index.html", courses=courses, cgpa=cgpa,
                            needed=needed, target=target, grade_options=grade_options)
 
 
 # Start the Flask development server
-# debug=True enables auto-reload on file save and shows detailed error pages
 if __name__ == "__main__":
     app.run(debug=True)
